move exceptions to common folder
