+++ api/common/exceptions.py
